+++ internal/pkg/service/get_avatar_settings.go
@@ -1,5 +1,5 @@
 package service
 
-func (s *mvmService) GetAvatarSettings(id string) (map[int64]int64, error) {
+func (s *mvmService) GetAvatarSettings(id string) (map[int32]int32, error) {
 	return s.store.GetAvatarSettings(id)
 }
