+++ internal/pkg/service/upsert_avatar_settings.go
@@ -1,5 +1,5 @@
 package service
 
-func (s *mvmService) UpsertAvatarSettings(userID string, settings map[int64]int64) error {
+func (s *mvmService) UpsertAvatarSettings(userID string, settings map[int32]int32) error {
 	return s.store.UpsertAvatarSettings(userID, settings)
 }
